Fix various issues and improve dream handling logic

+++ src/Business/Application/UseCases/Dreams/DreamDelete/DreamDeleteCommandHandler.cs
@@ -1,15 +1,22 @@
+using System.Text.Json;
+using System.Text.Json.Nodes;
+using System.Text.Json;
+using System.Text.Json.Nodes;
 using Application.DTO;
 using Application.Exceptions;
 using Domain.Entity;
 using Domain.IRepositories;
 using Domain.IService;
+using Domain.Model;
 using MediatR;
 
 namespace Application.UseCases.Dreams.DreamDelete;
 
 public class DreamDeleteCommandHandler (
     IUnitOfWork unitOfWork,
-    ICacheService<string, DreamResponseDto> cacheService
+    ICacheService<string, DreamResponseDto> cacheService,
+    ICacheService<DreamCacheKey, List<DreamResponseDto>> allDreamCacheService,
+    IHttpContextService httpContextService
     ): IRequestHandler<DreamDeleteCommand>
 {
     public async Task Handle(DreamDeleteCommand request, CancellationToken cancellationToken)
@@ -17,7 +24,12 @@ public class DreamDeleteCommandHandler (
         var dream = await unitOfWork.DreamRepository.GetAsync([request.DreamId], cancellationToken);
         if (dream is null) throw new NotFoundException("Dream not found.");
 
+        var currentUserId = httpContextService.GetCurrentUserId();
+        if (currentUserId != dream.ProducerId) 
+            throw new UnauthorizedException("You are not authorized to delete this dream.");
+
         await cacheService.RemoveAsync(request.DreamId.ToString() + nameof(Dream));
+        await allDreamCacheService.RemoveAsync(new DreamCacheKey());
 
         await unitOfWork.DreamRepository.DeleteAsync(dream, cancellationToken);
         await unitOfWork.SaveChangesAsync(cancellationToken);
